index.html: strikethrough old comment
The issue with the strange layout for some thumbnails ended up being
due to ImageMagick using the PDF's MediaBox by default. The MediaBox
is meant for print, and it is better for us to use the CropBox when
generating thumbnails because that defines the area meant to be dis-
played to a user on screen.

See: https://bugs.ghostscript.com/show_bug.cgi?id=705994